move constEvalutor to expr_simplififer
