+++ datafusion/optimizer/src/simplify_expressions/expr_simplifier.rs
@@ -18,12 +18,14 @@
 //! Expression simplification API
 
 use crate::{
-    simplify_expressions::{ConstEvaluator, Simplifier},
+    simplify_expressions::{Simplifier},
     type_coercion::TypeCoercionRewriter,
 };
 
-use datafusion_common::{DFSchemaRef, Result};
-use datafusion_expr::{expr_rewriter::ExprRewritable, Expr};
+use arrow::{record_batch::RecordBatch, datatypes::{Schema, DataType, Field}, array::new_null_array};
+use datafusion_common::{DFSchemaRef, Result, DFSchema, DataFusionError, ScalarValue};
+use datafusion_expr::{expr_rewriter::{ExprRewritable, ExprRewriter, RewriteRecursion}, Expr, Volatility, ColumnarValue};
+use datafusion_physical_expr::{execution_props::ExecutionProps, create_physical_expr};
 
 use super::SimplifyInfo;
 
@@ -126,16 +128,198 @@ impl<S: SimplifyInfo> ExprSimplifier<S> {
     }
 }
 
+
+#[allow(rustdoc::private_intra_doc_links)]
+/// Partially evaluate `Expr`s so constant subtrees are evaluated at plan time.
+///
+/// Note it does not handle algebraic rewrites such as `(a or false)`
+/// --> `a`, which is handled by [`Simplifier`]
+struct ConstEvaluator<'a> {
+    /// can_evaluate is used during the depth-first-search of the
+    /// Expr tree to track if any siblings (or their descendants) were
+    /// non evaluatable (e.g. had a column reference or volatile
+    /// function)
+    ///
+    /// Specifically, `can_evaluate[N]` represents the state of
+    /// traversal when we are N levels deep in the tree, one entry for
+    /// this Expr and each of its parents.
+    ///
+    /// After visiting all siblings if can_evauate.top() is true, that
+    /// means there were no non evaluatable siblings (or their
+    /// descendants) so this Expr can be evaluated
+    can_evaluate: Vec<bool>,
+
+    execution_props: &'a ExecutionProps,
+    input_schema: DFSchema,
+    input_batch: RecordBatch,
+}
+
+impl<'a> ExprRewriter for ConstEvaluator<'a> {
+    fn pre_visit(&mut self, expr: &Expr) -> Result<RewriteRecursion> {
+        // Default to being able to evaluate this node
+        self.can_evaluate.push(true);
+
+        // if this expr is not ok to evaluate, mark entire parent
+        // stack as not ok (as all parents have at least one child or
+        // descendant that can not be evaluated
+
+        if !Self::can_evaluate(expr) {
+            // walk back up stack, marking first parent that is not mutable
+            let parent_iter = self.can_evaluate.iter_mut().rev();
+            for p in parent_iter {
+                if !*p {
+                    // optimization: if we find an element on the
+                    // stack already marked, know all elements above are also marked
+                    break;
+                }
+                *p = false;
+            }
+        }
+
+        // NB: do not short circuit recursion even if we find a non
+        // evaluatable node (so we can fold other children, args to
+        // functions, etc)
+        Ok(RewriteRecursion::Continue)
+    }
+
+    fn mutate(&mut self, expr: Expr) -> Result<Expr> {
+        match self.can_evaluate.pop() {
+            Some(true) => Ok(Expr::Literal(self.evaluate_to_scalar(expr)?)),
+            Some(false) => Ok(expr),
+            _ => Err(DataFusionError::Internal(
+                "Failed to pop can_evaluate".to_string(),
+            )),
+        }
+    }
+}
+
+impl<'a> ConstEvaluator<'a> {
+    /// Create a new `ConstantEvaluator`. Session constants (such as
+    /// the time for `now()` are taken from the passed
+    /// `execution_props`.
+    pub fn try_new(execution_props: &'a ExecutionProps) -> Result<Self> {
+        // The dummy column name is unused and doesn't matter as only
+        // expressions without column references can be evaluated
+        static DUMMY_COL_NAME: &str = ".";
+        let schema = Schema::new(vec![Field::new(DUMMY_COL_NAME, DataType::Null, true)]);
+        let input_schema = DFSchema::try_from(schema.clone())?;
+        // Need a single "input" row to produce a single output row
+        let col = new_null_array(&DataType::Null, 1);
+        let input_batch = RecordBatch::try_new(std::sync::Arc::new(schema), vec![col])?;
+
+        Ok(Self {
+            can_evaluate: vec![],
+            execution_props,
+            input_schema,
+            input_batch,
+        })
+    }
+
+    /// Can a function of the specified volatility be evaluated?
+    fn volatility_ok(volatility: Volatility) -> bool {
+        match volatility {
+            Volatility::Immutable => true,
+            // Values for functions such as now() are taken from ExecutionProps
+            Volatility::Stable => true,
+            Volatility::Volatile => false,
+        }
+    }
+
+    /// Can the expression be evaluated at plan time, (assuming all of
+    /// its children can also be evaluated)?
+    fn can_evaluate(expr: &Expr) -> bool {
+        // check for reasons we can't evaluate this node
+        //
+        // NOTE all expr types are listed here so when new ones are
+        // added they can be checked for their ability to be evaluated
+        // at plan time
+        match expr {
+            // Has no runtime cost, but needed during planning
+            Expr::Alias(..)
+            | Expr::AggregateFunction { .. }
+            | Expr::AggregateUDF { .. }
+            | Expr::ScalarVariable(_, _)
+            | Expr::Column(_)
+            | Expr::Exists { .. }
+            | Expr::InSubquery { .. }
+            | Expr::ScalarSubquery(_)
+            | Expr::WindowFunction { .. }
+            | Expr::Sort { .. }
+            | Expr::GroupingSet(_)
+            | Expr::Wildcard
+            | Expr::QualifiedWildcard { .. } => false,
+            Expr::ScalarFunction { fun, .. } => Self::volatility_ok(fun.volatility()),
+            Expr::ScalarUDF { fun, .. } => Self::volatility_ok(fun.signature.volatility),
+            Expr::Literal(_)
+            | Expr::BinaryExpr { .. }
+            | Expr::Not(_)
+            | Expr::IsNotNull(_)
+            | Expr::IsNull(_)
+            | Expr::IsTrue(_)
+            | Expr::IsFalse(_)
+            | Expr::IsUnknown(_)
+            | Expr::IsNotTrue(_)
+            | Expr::IsNotFalse(_)
+            | Expr::IsNotUnknown(_)
+            | Expr::Negative(_)
+            | Expr::Between { .. }
+            | Expr::Like { .. }
+            | Expr::ILike { .. }
+            | Expr::SimilarTo { .. }
+            | Expr::Case(_)
+            | Expr::Cast { .. }
+            | Expr::TryCast { .. }
+            | Expr::InList { .. }
+            | Expr::GetIndexedField { .. } => true,
+        }
+    }
+
+    /// Internal helper to evaluates an Expr
+    pub(crate) fn evaluate_to_scalar(&mut self, expr: Expr) -> Result<ScalarValue> {
+        if let Expr::Literal(s) = expr {
+            return Ok(s);
+        }
+
+        let phys_expr = create_physical_expr(
+            &expr,
+            &self.input_schema,
+            &self.input_batch.schema(),
+            self.execution_props,
+        )?;
+        let col_val = phys_expr.evaluate(&self.input_batch)?;
+        match col_val {
+            ColumnarValue::Array(a) => {
+                if a.len() != 1 {
+                    Err(DataFusionError::Execution(format!(
+                        "Could not evaluate the expression, found a result of length {}",
+                        a.len()
+                    )))
+                } else {
+                    Ok(ScalarValue::try_from_array(&a, 0)?)
+                }
+            }
+            ColumnarValue::Scalar(s) => Ok(s),
+        }
+    }
+}
+
+
 #[cfg(test)]
 mod tests {
-    use crate::simplify_expressions::SimplifyContext;
+    use std::sync::Arc;
+
+    use crate::simplify_expressions::{SimplifyContext, utils::for_test::{now_expr, cast_to_int64_expr, to_timestamp_expr}};
 
     use super::*;
-    use arrow::datatypes::{Field, Schema, DataType};
+    use arrow::{datatypes::{Field, Schema, DataType}, array::{ArrayRef, Int32Array}};
+    use chrono::{TimeZone, DateTime, Utc};
     use datafusion_common::ToDFSchema;
-    use datafusion_expr::{col, lit, when};
-    use datafusion_physical_expr::execution_props::ExecutionProps;
+    use datafusion_expr::{col, lit, when, call_fn, BuiltinScalarFunction, lit_timestamp_nano, create_udf, ScalarUDF};
+    use datafusion_physical_expr::{execution_props::ExecutionProps, functions::make_scalar_function};
 
+    // ------------------------------
+    // --- ExprSimplifier tests -----
+    // ------------------------------
     #[test]
     fn api_basic() {
         let props = ExecutionProps::new();
@@ -210,5 +394,202 @@ mod tests {
             .unwrap();
         let expected = col("i").lt(lit(5));
         assert_eq!(expected, simplifier.simplify(expr).unwrap());
+    }
+
+
+    // ------------------------------
+    // --- ConstEvaluator tests -----
+    // ------------------------------
+    fn test_evaluate_with_start_time(
+        input_expr: Expr,
+        expected_expr: Expr,
+        date_time: &DateTime<Utc>,
+    ) {
+        let execution_props = ExecutionProps {
+            query_execution_start_time: *date_time,
+            var_providers: None,
+        };
+
+        let mut const_evaluator = ConstEvaluator::try_new(&execution_props).unwrap();
+        let evaluated_expr = input_expr
+            .clone()
+            .rewrite(&mut const_evaluator)
+            .expect("successfully evaluated");
+
+        assert_eq!(
+            evaluated_expr, expected_expr,
+            "Mismatch evaluating {}\n  Expected:{}\n  Got:{}",
+            input_expr, expected_expr, evaluated_expr
+        );
+    }
+
+    fn test_evaluate(input_expr: Expr, expected_expr: Expr) {
+        test_evaluate_with_start_time(input_expr, expected_expr, &Utc::now())
+    }
+
+    // Make a UDF that adds its two values together, with the specified volatility
+    fn make_udf_add(volatility: Volatility) -> Arc<ScalarUDF> {
+        let input_types = vec![DataType::Int32, DataType::Int32];
+        let return_type = Arc::new(DataType::Int32);
+
+        let fun = |args: &[ArrayRef]| {
+            let arg0 = &args[0]
+                .as_any()
+                .downcast_ref::<Int32Array>()
+                .expect("cast failed");
+            let arg1 = &args[1]
+                .as_any()
+                .downcast_ref::<Int32Array>()
+                .expect("cast failed");
+
+            // 2. perform the computation
+            let array = arg0
+                .iter()
+                .zip(arg1.iter())
+                .map(|args| {
+                    if let (Some(arg0), Some(arg1)) = args {
+                        Some(arg0 + arg1)
+                    } else {
+                        // one or both args were Null
+                        None
+                    }
+                })
+                .collect::<Int32Array>();
+
+            Ok(Arc::new(array) as ArrayRef)
+        };
+
+        let fun = make_scalar_function(fun);
+        Arc::new(create_udf(
+            "udf_add",
+            input_types,
+            return_type,
+            volatility,
+            fun,
+        ))
+    }
+
+    #[test]
+    fn test_const_evaluator() {
+        // true --> true
+        test_evaluate(lit(true), lit(true));
+        // true or true --> true
+        test_evaluate(lit(true).or(lit(true)), lit(true));
+        // true or false --> true
+        test_evaluate(lit(true).or(lit(false)), lit(true));
+
+        // "foo" == "foo" --> true
+        test_evaluate(lit("foo").eq(lit("foo")), lit(true));
+        // "foo" != "foo" --> false
+        test_evaluate(lit("foo").not_eq(lit("foo")), lit(false));
+
+        // c = 1 --> c = 1
+        test_evaluate(col("c").eq(lit(1)), col("c").eq(lit(1)));
+        // c = 1 + 2 --> c + 3
+        test_evaluate(col("c").eq(lit(1) + lit(2)), col("c").eq(lit(3)));
+        // (foo != foo) OR (c = 1) --> false OR (c = 1)
+        test_evaluate(
+            (lit("foo").not_eq(lit("foo"))).or(col("c").eq(lit(1))),
+            lit(false).or(col("c").eq(lit(1))),
+        );
+    }
+
+    #[test]
+    fn test_const_evaluator_scalar_functions() {
+        // concat("foo", "bar") --> "foobar"
+        let expr = call_fn("concat", vec![lit("foo"), lit("bar")]).unwrap();
+        test_evaluate(expr, lit("foobar"));
+
+        // ensure arguments are also constant folded
+        // concat("foo", concat("bar", "baz")) --> "foobarbaz"
+        let concat1 = call_fn("concat", vec![lit("bar"), lit("baz")]).unwrap();
+        let expr = call_fn("concat", vec![lit("foo"), concat1]).unwrap();
+        test_evaluate(expr, lit("foobarbaz"));
+
+        // Check non string arguments
+        // to_timestamp("2020-09-08T12:00:00+00:00") --> timestamp(1599566400000000000i64)
+        let expr =
+            call_fn("to_timestamp", vec![lit("2020-09-08T12:00:00+00:00")]).unwrap();
+        test_evaluate(expr, lit_timestamp_nano(1599566400000000000i64));
+
+        // check that non foldable arguments are folded
+        // to_timestamp(a) --> to_timestamp(a) [no rewrite possible]
+        let expr = call_fn("to_timestamp", vec![col("a")]).unwrap();
+        test_evaluate(expr.clone(), expr);
+
+        // check that non foldable arguments are folded
+        // to_timestamp(a) --> to_timestamp(a) [no rewrite possible]
+        let expr = call_fn("to_timestamp", vec![col("a")]).unwrap();
+        test_evaluate(expr.clone(), expr);
+
+        // volatile / stable functions should not be evaluated
+        // rand() + (1 + 2) --> rand() + 3
+        let fun = BuiltinScalarFunction::Random;
+        assert_eq!(fun.volatility(), Volatility::Volatile);
+        let rand = Expr::ScalarFunction { args: vec![], fun };
+        let expr = rand.clone() + (lit(1) + lit(2));
+        let expected = rand + lit(3);
+        test_evaluate(expr, expected);
+
+        // parenthesization matters: can't rewrite
+        // (rand() + 1) + 2 --> (rand() + 1) + 2)
+        let fun = BuiltinScalarFunction::Random;
+        let rand = Expr::ScalarFunction { args: vec![], fun };
+        let expr = (rand + lit(1)) + lit(2);
+        test_evaluate(expr.clone(), expr);
+    }
+
+    #[test]
+    fn test_const_evaluator_now() {
+        let ts_nanos = 1599566400000000000i64;
+        let time = chrono::Utc.timestamp_nanos(ts_nanos);
+        let ts_string = "2020-09-08T12:05:00+00:00";
+        // now() --> ts
+        test_evaluate_with_start_time(now_expr(), lit_timestamp_nano(ts_nanos), &time);
+
+        // CAST(now() as int64) + 100_i64 --> ts + 100_i64
+        let expr = cast_to_int64_expr(now_expr()) + lit(100_i64);
+        test_evaluate_with_start_time(expr, lit(ts_nanos + 100), &time);
+
+        //  CAST(now() as int64) < cast(to_timestamp(...) as int64) + 50000_i64 ---> true
+        let expr = cast_to_int64_expr(now_expr())
+            .lt(cast_to_int64_expr(to_timestamp_expr(ts_string)) + lit(50000i64));
+        test_evaluate_with_start_time(expr, lit(true), &time);
+    }
+
+    #[test]
+    fn test_evaluator_udfs() {
+        let args = vec![lit(1) + lit(2), lit(30) + lit(40)];
+        let folded_args = vec![lit(3), lit(70)];
+
+        // immutable UDF should get folded
+        // udf_add(1+2, 30+40) --> 73
+        let expr = Expr::ScalarUDF {
+            args: args.clone(),
+            fun: make_udf_add(Volatility::Immutable),
+        };
+        test_evaluate(expr, lit(73));
+
+        // stable UDF should be entirely folded
+        // udf_add(1+2, 30+40) --> 73
+        let fun = make_udf_add(Volatility::Stable);
+        let expr = Expr::ScalarUDF {
+            args: args.clone(),
+            fun: Arc::clone(&fun),
+        };
+        test_evaluate(expr, lit(73));
+
+        // volatile UDF should have args folded
+        // udf_add(1+2, 30+40) --> udf_add(3, 70)
+        let fun = make_udf_add(Volatility::Volatile);
+        let expr = Expr::ScalarUDF {
+            args,
+            fun: Arc::clone(&fun),
+        };
+        let expected_expr = Expr::ScalarUDF {
+            args: folded_args,
+            fun: Arc::clone(&fun),
+        };
+        test_evaluate(expr, expected_expr);
     }
 }
