+++ datafusion/optimizer/src/simplify_expressions/utils.rs
@@ -352,3 +352,24 @@ pub fn simpl_concat_ws(delimiter: &Expr, args: &[Expr]) -> Result<Expr> {
         )),
     }
 }
+
+#[cfg(test)]
+pub mod for_test {
+    use arrow::datatypes::DataType;
+    use datafusion_expr::{Expr, call_fn, lit};
+
+    pub fn now_expr() -> Expr {
+        call_fn("now", vec![]).unwrap()
+    }
+
+    pub fn cast_to_int64_expr(expr: Expr) -> Expr {
+        Expr::Cast {
+            expr: expr.into(),
+            data_type: DataType::Int64,
+        }
+    }
+
+    pub fn to_timestamp_expr(arg: impl Into<String>) -> Expr {
+        call_fn("to_timestamp", vec![lit(arg.into())]).unwrap()
+    }
+}
